src/main/java/org/squashleague/web/interceptor/navigation/NavigationInterceptor.java
package org.squashleague.web.interceptor.navigation;

import com.google.common.base.Function;
import com.google.common.base.Strings;
import com.google.common.collect.Lists;
import org.apache.commons.lang3.StringUtils;
import org.springframework.web.servlet.ModelAndView;
import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
import org.squashleague.web.interceptor.bundling.WroModelHolder;
import ro.isdc.wro.model.WroModel;
import ro.isdc.wro.model.group.Group;
import ro.isdc.wro.model.resource.Resource;
import ro.isdc.wro.model.resource.ResourceType;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.util.*;

/**
 * @author jamesdbloom
 */
public class NavigationInterceptor extends HandlerInterceptorAdapter {

    @Override
    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
        if (!modelAndView.getViewName().startsWith("redirect:")) {
            modelAndView.addObject("page", StringUtils.substringAfterLast(request.getRequestURI(), "/"));
        }
    }

}
